rename tab e2e files to match route path
